builder/interface.ts and builder/schema-to-ts.ts are now together entirely self contained!
