+++ packages/builder/src/index.ts
@@ -1,20 +1,6 @@
 // Re-export all public API from interface for backward compatibility
 export * from "./interface.ts";
 
-// Re-export builders with actual implementations for backward compatibility
-export { opaqueRef as cell, stream } from "./opaque-ref.ts";
-export {
-  byRef,
-  compute,
-  createNodeFactory,
-  derive,
-  handler,
-  lift,
-  render,
-} from "./module.ts";
-export {
-  recipe,
-} from "./recipe.ts";
 export {
   type BuiltInCompileAndRunParams,
   type BuiltInCompileAndRunState,
@@ -74,12 +60,3 @@ export {
 // Export the factory function
 export { createBuilder } from "./factory.ts";
 export type { BuilderFunctions, BuilderRuntime } from "./interface.ts";
-
-// Export createCell function placeholder
-export function createCell(...args: any[]): any {
-  throw new Error(
-    "createCell requires a runtime context. " +
-      "This function should be overridden during recipe evaluation. " +
-      "If you're seeing this error, it means createCell is being called outside of a recipe context.",
-  );
-}
